Translate part of segment Segment_1d12 and fix Index out of range bugs

+++ src/CivGame/CivGameGlobals.cs
@@ -81,7 +81,7 @@ namespace OpenCiv1
 		public ushort Var_b2ba = 0;
 		public ushort Var_d2de = 0;
 		public int Var_d2e0 = 0;
-		public ushort Var_d2f6 = 0;
+		public int Var_d2f6 = 0;
 		public short Var_d4cc_XPos = 0;
 		public short Var_d75e_YPos = 0;
 		public ushort Var_d762 = 0;
@@ -89,10 +89,10 @@ namespace OpenCiv1
 		public ushort Var_db3a = 0;
 		public ushort Var_db3c = 0;
 		public ushort Var_db3e = 0;
-		public ushort Var_db42 = 0;
-		public ushort Var_deb8 = 0;
+		public int Var_db42 = 0;
+		public int Var_deb8 = 0;
 		public int Var_e17a = 0;
-		public ushort Var_e3c2 = 0;
+		public int Var_e3c2 = 0;
 		public ushort Var_e3c6 = 0;
 		public int Var_e8b8 = 0;
 		#endregion
